rename walletconnectv2 to walletconnect
